requirement.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
django-simple-serializer==2.0.7
djangorestframework==3.7.1
future==0.16.0
gunicorn==19.7.1
idna==2.6
mysqlclient==1.3.12
pytz==2017.2
requests==2.18.4
urllib3==1.22
